Added a print self.__str__() to show winning state
The winning state was not showng at the end of the game. The only thing that was printed was the last step someone took So I added the print(self.__str__()) after the main while loop to print the winning state. I also added a new line after game tied or game won is mentioned.